chore: moved timeline viewer to new component
